pyproject.toml
[tool.poetry]
name = "crypto-swarm"
version = "0.1.0"
description = "Phase 0.1 setup for crypto trading swarm."
authors = ["Patrick"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pyyaml = "^6.0.2"
asyncpg = "^0.30.0"
redis = "^6.0.0"
httpx = "^0.28.1"
ccxt = "^4.4.84"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^1.0.0"
ruff = "^0.11.8"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
